Add files via upload
All commands seem to work. We need to find a way to print out the received directories from ls. Everything else works. However, when we use a passive command and want to go use a command that is not passive, the program seems to crash. We are not using the socket with the remote port and we suspect that may have to do something with it.  We also changed the input so it could match the project guidelines.